refactor: add keyBoardBuilder
